+++ src/Queue/src/Exception/FailException.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Spiral\Queue\Exception;
+
+final class FailException extends StateException
+{
+    public function __construct(string $reason = '')
+    {
+        parent::__construct($reason);
+    }
+}